+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Generally, more significant breaking changes will be put near the top of each category.
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed incorrectly written test `test_maybe.test_maybe_or`
+
 ## [0.8.0]
 
 ### Added